mediapipe-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
mediapipe==0.10.18
opencv-python-headless==4.10.0.84
numpy==1.26.4
ffmpeg-python==0.2.0
pydantic==2.9.2
httpx==0.27.2
